Refactoring of sensu-backend entry point
