+++ backend/cmd/root.go
@@ -1,24 +1,19 @@
 package cmd
 
-import (
-	"github.com/sensu/sensu-go/backend"
-	"github.com/spf13/cobra"
-)
+// // initializeFunc represents the signature of an initialization function, used
+// // to initialize the backend
+// type initializeFunc func(*backend.Config) (*backend.Backend, error)
 
-// initializeFunc represents the signature of an initialization function, used
-// to initialize the backend
-type initializeFunc func(*backend.Config) (*backend.Backend, error)
+// var (
+// 	initialize initializeFunc
+// 	rootCmd    = &cobra.Command{
+// 		Use:   "sensu-backend",
+// 		Short: "sensu backend",
+// 	}
+// )
 
-var (
-	initialize initializeFunc
-	rootCmd    = &cobra.Command{
-		Use:   "sensu-backend",
-		Short: "sensu backend",
-	}
-)
-
-// Execute ...
-func Execute(fn initializeFunc) error {
-	initialize = fn
-	return rootCmd.Execute()
-}
+// // Execute ...
+// func Execute(fn initializeFunc) error {
+// 	initialize = fn
+// 	return rootCmd.Execute()
+// }
