--- cmd/sensu-backend/logger.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/sirupsen/logrus"
-
-var logger = logrus.WithFields(logrus.Fields{
-	"component": "backend",
-})
